Refactor around function
Now it returns all valid cells around it and the caller is meant to iterate over them rather than passing an action. It also uses reduce and generalizes how it checks each cell.
This also finally uses the around function to populate the count values for the first row of sections.